requirements.txt
anyio==3.5.0
asgiref==3.5.1
click==8.1.3
colorama==0.4.4
fastapi==0.77.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
pydantic==1.9.0
python-dotenv==0.20.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
